🛡️ Sentinel: [CRITICAL] Fix data loss on corrupt deserialization
Refactored `TimerRepository` to use a dedicated `TimerSerializer` that enforces strict deserialization (throwing exceptions) during write operations. This prevents the application from silently reading corrupted data as an empty list and overwriting it, effectively resolving a potential data loss vulnerability in "read-modify-write" cycles.

Key changes:
- Extracted JSON logic to `TimerSerializer` in `core` module.
- Added `throwOnFailure` parameter to deserialization methods.
- Enabled strict mode for `save`/`delete`/`add`/`remove` operations in `TimerRepository`.
- Kept recovery mode (fail-safe) for UI `Flow` reads.
- Added unit tests for `TimerSerializer`.
- Fixed syntax error in `WeatherAssistantCard.kt`.
- Updated `core/build.gradle.kts` to support unit testing of Android logic.
