+++ app/src/main/kotlin/com/dagsbalken/app/ui/dagskompisen/WeatherAssistantCard.kt
@@ -1,7 +1,7 @@
 package com.dagsbalken.app.ui.dagskompisen
 
 import androidx.compose.foundation.Image
-§§import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
