+++ core/build.gradle.kts
@@ -28,6 +28,10 @@ android {
         sourceCompatibility = JavaVersion.VERSION_17
         targetCompatibility = JavaVersion.VERSION_17
     }
+
+    testOptions {
+        unitTests.isReturnDefaultValues = true
+    }
 }
 
 kotlin {
@@ -44,4 +48,5 @@ dependencies {
 
     // Unit testing
     testImplementation("junit:junit:4.13.2")
+    testImplementation("org.json:json:20231013")
 }
